Hygeine activity: Refactored Student service class.
